+++ bes_ml2/elm_trainer.py
@@ -7,8 +7,9 @@ from pytorch_lightning import callbacks as cb
 from pytorch_lightning.utilities.model_summary import ModelSummary
 import wandb
 
-import elm_data
-import elm_model
+import elm_datamodule
+import elm_lightning_model
+import elm_torch_model
 
 
 @dataclasses.dataclass(eq=False)
@@ -22,22 +23,23 @@ class BES_Trainer:
     early_stopping_patience: int = 10
     enable_progress_bar: bool = True
     num_nodes: int = 1
-    datamodule: pl.LightningDataModule = None
-    lit_model: pl.LightningModule = None
-    monitor: str = 'val_score'
+    datamodule: elm_datamodule.ELM_Datamodule = None
+    lightning_model: elm_lightning_model.Lightning_Model = None
+    monitor_metric: str = 'val_score'
     wandb_log_freq: int = 100
     pl_log_freq: int = 50
 
     def __post_init__(self):
-        assert self.datamodule and self.lit_model
-        if self.wandb_log:
-            wandb.login()
+        assert self.datamodule and self.lightning_model
+        assert self.datamodule.signal_window_size == self.lightning_model.signal_window_size
+
+        self.monitor_metric = self.lightning_model.monitor_metric
         self.trainer = None
 
         print(f'Initiating {self.__class__.__name__}')
         class_fields_dict = {field.name: field for field in dataclasses.fields(self.__class__)}
         for field_name in dataclasses.asdict(self):
-            if field_name in ['datamodule', 'model']:
+            if field_name in ['datamodule', 'lit_model']:
                 continue
             value = getattr(self, field_name)
             field_str = f"  {field_name}: {value}"
@@ -47,7 +49,7 @@ class BES_Trainer:
             print(field_str)
 
         print("Model Summary:")
-        print(ModelSummary(self.lit_model, max_depth=-1))
+        print(ModelSummary(self.lightning_model, max_depth=-1))
 
     def run_fast_dev(self):
         tmp_trainer = pl.Trainer(
@@ -56,7 +58,7 @@ class BES_Trainer:
             enable_model_summary=False,
         )
         tmp_trainer.fit(
-            model=self.lit_model, 
+            model=self.lightning_model, 
             datamodule=self.datamodule,
         )
 
@@ -73,17 +75,18 @@ class BES_Trainer:
         )
         self.loggers.append(tb_logger)
         version_str = tb_logger.version if tb_logger.version is str else f"version_{tb_logger.version}"
-        if hasattr(self.lit_model, 'log_dir'):
-            self.lit_model.log_dir = tb_logger.log_dir
+        if hasattr(self.lightning_model, 'log_dir'):
+            self.lightning_model.log_dir = tb_logger.log_dir
         
         if self.wandb_log:
+            wandb.login()
             wandb_logger = loggers.WandbLogger(
                 save_dir=self.experiment_group_dir,
                 project=os.path.basename(self.experiment_group_dir),
                 name=version_str,
             )
             wandb_logger.watch(
-                self.lit_model, 
+                self.lightning_model, 
                 log='all', 
                 log_freq=self.wandb_log_freq,
             )
@@ -93,8 +96,8 @@ class BES_Trainer:
         self.callbacks = [
             cb.LearningRateMonitor(),
             cb.EarlyStopping(
-                monitor=self.monitor,
-                mode='min' if 'loss' in self.monitor else 'max',
+                monitor=self.monitor_metric,
+                mode='min' if 'loss' in self.monitor_metric else 'max',
                 min_delta=self.early_stopping_min_delta,
                 patience=self.early_stopping_patience,
             ),
@@ -116,14 +119,18 @@ class BES_Trainer:
             devices="auto",
         )
 
-        self.trainer.fit(self.lit_model, datamodule=self.datamodule)
+        self.trainer.fit(self.lightning_model, datamodule=self.datamodule)
         self.trainer.test(datamodule=self.datamodule, ckpt_path='best')
         self.trainer.predict(datamodule=self.datamodule, ckpt_path='best')
 
 
 if __name__=='__main__':
     signal_window_size = 512
-    datamodule = elm_data.ELM_Datamodule(
+
+    """
+    Step 1a: Initiate pytorch_lightning.LightningDataModule
+    """
+    datamodule = elm_datamodule.ELM_Datamodule(
         # data_file='/global/homes/d/drsmith/ml/scratch/data/labeled_elm_events.hdf5',
         signal_window_size=signal_window_size,
         max_elms=100,
@@ -131,17 +138,27 @@ if __name__=='__main__':
         fraction_validation=0.1,
         fraction_test=0.1,
     )
-    lit_model = elm_model.Lit_Model(
+
+    """
+    Step 1b: Initiate torch and lightning models
+        Ugly hack: must initiate lightning model, then initiate torch model, 
+        then add torch model to lightning model
+    """
+    lightning_model = elm_lightning_model.Lightning_Model(
         lr=1e-3,
         weight_decay=1e-5,
     )
-    torch_model = elm_model.Torch_Model_CNN02(
+    torch_model = elm_torch_model.Torch_Model_CNN02(
         signal_window_size=signal_window_size,
-        mlp_layers_size=(64,32,32),
+        mlp_layers=(64,32,32),
     )
-    lit_model.set_torch_model(torch_model=torch_model)
+    lightning_model.set_torch_model(torch_model=torch_model)
+
+    """
+    Step 2: Initiate pytorch_lightning.Trainer and run
+    """
     trainer = BES_Trainer(
-        lit_model=lit_model,
+        lightning_model=lightning_model,
         datamodule=datamodule,
         max_epochs=2,
         wandb_log=False,
